pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
src_paths = ["api"]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 160
multi_line_output = 3
use_parentheses = true

[tool.black]
line-length = 160

[tool.pylint.main]
init-hook="import sys; import os; sys.path.append(os.path.join(os.getcwd(), 'api'))"
ignore-paths = ["api/tests", "api/db_migrations"]

[tool.pylint.design]
max-args = 6

[tool.pylint."messages control"]
max-line-length = 160
disable = [
    "invalid-name",
    "missing-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-public-methods",
    "too-many-return-statements",
    "too-many-statements",
    "wrong-import-position",
    "unused-argument",
    "import-outside-toplevel",
    "broad-exception-caught",
    "broad-exception-raised",
    "too-many-positional-arguments",
    "too-many-arguments",
    "duplicate-code"
]

[tool.pylint.'FORMAT']
max-branches = 15
max-line-length = 160
max-statements = 60

[tool.pytest.ini_options]
addopts = "-v --cov api/ --cov-report term-missing --cov-fail-under 80 --log-level DEBUG --ignore=api/tests/api --import-mode=importlib"
norecursedirs = ["api/db_migrations", ".git", ".tox", "node_modules", "__pycache__", "api/tests/api"]
testpaths = ["api/tests/unit"]
pythonpath = ["api"]
markers = [
    "functional: marks tests as functional tests requiring Docker (deselect with '-m \"not functional\"')"
]

[tool.poetry]
name = "brewhouse-manager"
version = "0.7.1"
authors = ["alanquillin"]
description = "Brewhouse Manager"

[tool.poetry.dependencies]
MarkupSafe = "2.1.1"
cryptography = "46.0.3"
psycopg2-binary = "2.9.11"
fastapi = "^0.128.0"
PyJWT = "^2.10.1"
SQLAlchemy-Utils = "^0.42.1"
Werkzeug = "^3.1.5"
alembic = "^1.6.5"
nanoid = "^2.0.0"
pgpy = "^0.6.0"
python = ">=3.10,<4.0"
python-dateutil = "^2.8.0"
requests = "^2.24.0"
schema = "^0.7.3"
simplejson = "^3.17.2"
six = "^1.15.0"
urllib3 = "^2.6.3"
oauthlib = "^3.1.1"
pyOpenSSL = "^25.3.0"
argon2-cffi = "^25.1.0"
boto3 = "^1.37.7"
asyncpg = "^0.31.0"
python-multipart = "^0.0.22"
itsdangerous = "^2.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
uvicorn = {extras = ["standard"], version = "^0.40.0"}
httpx = "^0.28.1"
google-api-python-client = "^2.188.0"
google-auth-oauthlib = "^1.2.4"


[tool.poetry.dev-dependencies]
bandit = "^1.9.3"
black = "^26.1.0"
freezegun = "^1.5.5"
isort = "^7.0.0"
pdbpp = "^0.11.7"
py = "^1.10.0"
pylint = "^4.0.4"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-docker-tools = "^3.1.9"
requests_mock = "^1.8.0"
responses = "<0.25.8" # remove this one responses fixes their bug
wrapt = "^2.1.1" # required by pylint (badly)